add page preview quiz

+++ moodle_mobile/lib/models/quiz/quizData.dart
@@ -1,7 +1,7 @@
 import 'package:moodle_mobile/models/quiz/question.dart';
 
 class QuizData {
-  int? grade;
+  double? grade;
   AttemptData? attempt;
   List<Question>? questions;
